refactor(filid): split core/analysis + coverage-verify + pr-summary + remaining cache-manager
Batch C: 9 files split into single-fn files with facade preservation.

- core/analysis/dependency-graph (4 fn) → 4 files
- core/analysis/lca-calculator (3 fn) → 3 files
- core/analysis/project-analyzer (3 fn) → 2 files
- core/analysis/project-analyzer/renderers/report-renderer (3 fn) → 2 files
- core/coverage-verify/usage-tracker (3 fn) → 2 files
- core/pr-summary/aggregators/resolve-verdict (3 fn) → 2 files
- core/infra/cache-manager/caches/fractal-map-cache (3 fn) → 3 files
- core/infra/cache-manager/caches/prompt-context-cache (3 fn) → 3 files
- ast/parser (3 fn) → 3 files
