+++ packages/filid/src/core/coverage-verify/usage-tracker/get-module-name.ts
@@ -1,0 +1,10 @@
+import { basename } from 'node:path';
+
+/**
+ * Get the basename of a file without extension.
+ * Useful for test file matching.
+ */
+export function getModuleName(filePath: string): string {
+  const base = basename(filePath);
+  return base.replace(/\.[^.]+$/, '');
+}
