+++ packages/filid/src/core/pr-summary/aggregators/extract-verdict.ts
@@ -1,0 +1,7 @@
+/** review-report.md에서 verdict를 추출한다. */
+export function extractVerdict(content: string): string {
+  const match = content.match(
+    /\*\*Verdict\*\*:\s*(APPROVED|REQUEST_CHANGES|INCONCLUSIVE)/,
+  );
+  return match?.[1] ?? 'UNKNOWN';
+}
