+++ packages/filid/src/core/pr-summary/aggregators/resolve-verdict.ts
@@ -1,24 +1,8 @@
-/** review-report.md에서 verdict를 추출한다. */
-function extractVerdict(content: string): string {
-  const match = content.match(
-    /\*\*Verdict\*\*:\s*(APPROVED|REQUEST_CHANGES|INCONCLUSIVE)/,
-  );
-  return match?.[1] ?? 'UNKNOWN';
-}
+export { extractVerdict } from './extract-verdict.js';
+export { extractRevalidateVerdict } from './extract-revalidate-verdict.js';
 
-/** re-validate.md에서 verdict를 추출한다. */
-function extractRevalidateVerdict(content: string): string {
-  const headerMatch = content.match(/—\s*(PASS|FAIL)/);
-  if (headerMatch) return headerMatch[1];
-
-  const verdictMatch = content.match(/\*\*Verdict\*\*:\s*(PASS|FAIL)/);
-  if (verdictMatch) return verdictMatch[1];
-
-  const finalMatch = content.match(/\*\*Final Verdict\*\*:\s*(PASS|FAIL)/);
-  if (finalMatch) return finalMatch[1];
-
-  return 'UNKNOWN';
-}
+import { extractVerdict } from './extract-verdict.js';
+import { extractRevalidateVerdict } from './extract-revalidate-verdict.js';
 
 /** review-report.md와 re-validate.md에서 최종 verdict를 결정한다. */
 export function resolveVerdict(
